requirements.txt
python==3.12.4
audio_recorder_streamlit==0.0.10
baml_py==0.57.1
httpx_oauth==0.15.1
langchain==0.3.1
langchain_community==0.3.1
langchain_huggingface==0.1.0
langchain_openai==0.2.1
mediapipe==0.10.15
numpy==2.1.1
openai==1.51.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
pdfplumber==0.11.4
pinecone==5.3.1
pinecone_text==0.9.0
playsound==1.3.0
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
SpeechRecognition==3.10.4
streamlit==1.38.0
streamlit_on_Hover_tabs==1.0.1
typing_extensions==4.12.2
